App not fully running but fragment is implemented.
