+++ app/build.gradle
@@ -5,7 +5,7 @@ buildscript {
 }
 
 plugins {
-    id "io.sentry.android.gradle" version "2.1.0"
+    id "io.sentry.android.gradle"
 }
 
 apply plugin: 'com.android.application'
@@ -79,6 +79,7 @@ dependencies {
     implementation 'androidx.navigation:navigation-ui:2.1.0'
     implementation 'com.android.support:support-core-utils:28.0.0'
     implementation 'com.squareup.okhttp3:okhttp:4.9.0'
+    implementation 'androidx.fragment:fragment:1.3.2'
 }
 
 sentry {
