+++ sentry.properties
@@ -1,4 +1,4 @@
 # how a sentry.properties looks like
 defaults.org=testorg-az
 defaults.project=androidadamtest
-auth.token=2e3d1fa6c35043429509932efc2730a712ee1211c0764e9b868b6ad816f3948c
+auth.token=
